Disable pdf.js and prompt the user for action
With Firefox 29, pdf.js is no longer functional.  This
disables it, and disables full-page PDF plugins, so the
user is asked what to do with a PDF. (Trac: #1466)

A likely side-effect of this is that people who have configured the
Adobe plugin (or another PDF plugin) will now get prompted with the
dialog box, but we believe that to be preferable to printing not
working at all.