Remove last vestiges of autoconf; add VERSION
The only reason this is still autoconf is to support --prefix and
to work on platforms that don't have a working install(1) or a
mkdir -p.  We don't care about those platforms.

Convert Makefile.in to a trivial Makefile, and add a VERSION file
for use with the new build system